separated note/notes/newnote
