+++ src/client/features/plants/Note.jsx
@@ -1,49 +1,20 @@
-import { useGetNoteQuery, 
-    useGetNotesQuery, 
-    useCreateNoteMutation, 
-    useEditNoteMutation, 
-    useDeleteNoteMutation } from "../../store/Notes";
+import { useState } from "react";
+import { useEditNoteMutation, useDeleteNoteMutation } from "../../store/Notes";
 
-import { useSelector } from "react-redux";
-import { selectToken } from "../auth/authSlice";
 
-//~ function for rendering our createNoteForm  // window to pop up on screen
-function createNoteForm({ note }) {
-    //use the hooks to create the mutation functions we will use
-    //const [saveCreatedNote] = useSaveCreatedNoteMutation()
+export default function Note ({ note }){
 
-    const closePopupWindow = () => {
-        // "x" button for closing window
-        ///onclick close window
-        ////error if text!=== 0 {return "Are you sure you want to close without saving note?"}
-    }
-    //const saveCreatedNote = () => {
-        // "Save Note" button to add 
-        ///onclick save note and close window
-        ////error if text === 0 {return "There is no note to save"}
-    //}
-
-    return (
-        <>
-            <form>
-                <textarea name="note" rows="10" cols="20" />
-                <button onClick={closePopupWindow} id="delete">X</button>
-                <button onClick={saveCreatedNote} id="submit">Create Note</button>
-            </form>
-        </>
-    )
-}
 
 
 //~ function for rendering our editNoteForm  // window to pop up on screen
-function editNoteForm() {
-    //use the hooks to create the mutation functions we will use
-    //const [saveChangedNote] = useSaveChangedNoteMutation()
+// function editNoteForm() {
+//     //use the hooks to create the mutation functions we will use
+//     //const [saveChangedNote] = useSaveChangedNoteMutation()
 
-    const closePopupWindow = () => {  // "x" button for closing window
-        ///onclick close window
-        ////error if no change in text {return "No changes are saved"}
-    }
+//     const closePopupWindow = () => {  // "x" button for closing window
+//         ///onclick close window
+//         ////error if no change in text {return "No changes are saved"}
+//     }
     //const saveChangedNote = () => {
         // input field for text
         // text.value = text[selected note id] ; This will be the previous saved Note
@@ -51,16 +22,16 @@ function editNoteForm() {
         ///onclick update state, delete previous data and close popup window
     //}
 
-    return (
-        <>
-            <form>
-                <textarea name="note" rows="10" cols="20" />
-                <button onClick={closePopupWindow} id="delete">X</button>
-                <button onClick={saveChangedNote} id="submit">Create Note</button>
-            </form>
-        </>
-    )
-};
+//     return (
+//         <>
+//             <form>
+//                 <textarea name="note" rows="10" cols="20" />
+//                 <button onClick={closePopupWindow} id="delete">X</button>
+//                 <button onClick={saveChangedNote} id="submit">Create Note</button>
+//             </form>
+//         </>
+//     )
+// };
 
 
 //getElement by id createNoteBTN
@@ -68,28 +39,4 @@ function editNoteForm() {
 //getElement by id editBTN
 //onclick editBTN(editNoteForm)
 
-export default function Notes(){
-const token = useSelector(selectToken);
-const { data: notes, isLoading } = useGetNotesQuery();
-
-console.log(notes);
-
-if (!token) {
-    return <p>You must be logged in to see your tasks.</p>;
-};
-
-
-    return (
-    <div className="notes">
-        <h2>My Plant Notes</h2>
-        {isLoading && <p>Loading notes...</p>}
-        {notes && (
-            <ul>
-                {notes.map((note) => (
-                    <Notes key={note.id} note={note} />
-                ))}
-            </ul>
-        )}
-    </div>
-    );
-};
+}
